用户服务独立

+++ goya-issue/goya-issue-service/src/main/java/com/goya/issue/service/repository/IssuePriorityRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IssuePriorityRepository extends BaseRepository<IssuePriority, Integer> {
+public interface IssuePriorityRepository extends BaseRepository<IssuePriority, Long> {
 }
